chore: split erc20 into it's own file
